debuging the training ppl issue;
